fix 64bit warnings
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@4240 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ regutil.cpp
@@ -157,7 +157,7 @@ HRESULT SaveValueString(const std::string &szKey, const std::string &szValue, co
 
 HRESULT SaveValueString(const std::string &szKey, const std::string &szValue, const string& val)
 {
-   return SaveValue(szKey, szValue, REG_SZ, val.c_str(), val.length());
+   return SaveValue(szKey, szValue, REG_SZ, val.c_str(), (DWORD)val.length());
 }
 
 HRESULT DeleteValue(const std::string &szKey, const std::string &szValue)
